Add external file for core_id calculations
