+++ src/lib.rs
@@ -8,3 +8,4 @@ extern crate timely_logging as logging_core;
 pub mod execute;
 pub mod allocator;
 pub mod initialize;
+pub mod core_id;
